Se agrega nueva pestaña para el registro e inicio de sesion, ademas se da funcionalidad al boton inicio sesion

+++ src/main/java/ingreso/com/libreMercado/model/DaoUsuario.java
@@ -2,5 +2,6 @@ package ingreso.com.libreMercado.model;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface DaoUsuario extends CrudRepository<Usuario, Integer> {
+public interface DaoUsuario extends CrudRepository<Usuario, Long> {
+
 }
